% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{connect}
\alias{connect}
\title{Connects to an RDBMS.}
\usage{
connect(drv, ...)
}
\arguments{
\item{drv}{The \code{DBI} driver to use.}

\item{...}{Additional parameters to pass to the underlying
\code{DBI::dbConnect}}
}
\value{
\code{DBI} connection.
}
\description{
Note that this function returns available valid connection if
there exists any instead of creating one. If the connection does
not exist, a new one is created and saved globally in
\code{options} before being returned.
}
\examples{
## Establish a connection:
connection <- connect(RSQLite::SQLite(), ":memory:")

## The connection is the same as the one saved globally:
identical(connection, getConnection())

## Disconnect:
disconnect()

}

