% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{existsRow}
\alias{existsRow}
\title{Checks if any rows exist.}
\usage{
existsRow(.table, ...)
}
\arguments{
\item{.table}{The name of the database table.}

\item{...}{Where clause to be \code{AND}ed.}
}
\value{
Boolean indicating if any rows exist.
}
\description{
Checks if any rows exist.
}

