% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{getConnection}
\alias{getConnection}
\title{Returns existing connection.}
\usage{
getConnection(validate = TRUE)
}
\arguments{
\item{validate}{Indicates if the connection should be validated.}
}
\value{
The connection.
}
\description{
If validation is enabled, the function will stop on invalid
connection encounter.
}

