% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{insertRow}
\alias{insertRow}
\title{Insert a given set of values as rows.}
\usage{
insertRow(.table, ...)
}
\arguments{
\item{.table}{The name of the database table.}

\item{...}{Values}
}
\value{
Nothing special.
}
\description{
Insert a given set of values as rows.
}

