% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{readTable}
\alias{readTable}
\title{Reads a table from the database.}
\usage{
readTable(name)
}
\arguments{
\item{name}{The name of the database table.}
}
\value{
The database table contents.
}
\description{
Reads a table from the database.
}

