% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{selectRow}
\alias{selectRow}
\title{Selects rows.}
\usage{
selectRow(.table, ..., .ordering = NULL)
}
\arguments{
\item{.table}{The name of the database table.}

\item{...}{Where clause to be \code{AND}ed.}

\item{.ordering}{Ordering if any.}
}
\value{
Result.
}
\description{
Selects rows.
}

