% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{upsertRow}
\alias{upsertRow}
\title{Upserts row.}
\usage{
upsertRow(.table, .where, ...)
}
\arguments{
\item{.table}{The name of the table.}

\item{.where}{The where clause.}

\item{...}{Values to be updated.}
}
\value{
Nothing special.
}
\description{
Upserts row.
}

