% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{writeTable}
\alias{writeTable}
\title{Creates and/or populates a database table.}
\usage{
writeTable(.table, x, append = TRUE, overwrite = FALSE,
  withRowNames = FALSE)
}
\arguments{
\item{.table}{The name of the database table.}

\item{x}{The data to be populated. If contents are empty, it is
used to introspect the database structure to create a table if
it does not exist.}

\item{append}{If the values will be append. Otherwise will raise
exception if the table is not new and \code{overwrite} is
\code{FALSE}.}

\item{overwrite}{If the values will be overridden. Otherwise will
raise exception if the table is not new and \code{append} is
\code{FALSE}.}

\item{withRowNames}{Indicates if we are we inserting with row names.}
}
\value{
The success indicator as a logical value.
}
\description{
Creates and/or populates a database table.
}

