% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{writeData}
\alias{writeData}
\title{Write the data to the database table.}
\usage{
writeData(.table, x)
}
\arguments{
\item{.table}{The database table name.}

\item{x}{The data}
}
\value{
Nothing interesting
}
\description{
This allows partial data insertion.
}

